move button into widgets, added font-family to .plotly-editor
