+++ src/styles/components/_main.scss
@@ -2,4 +2,3 @@
 @import 'containers/main';
 @import 'fields/main';
 @import 'widgets/main';
-@import "button";
