+++ src/styles/components/widgets/_main.scss
@@ -1,3 +1,4 @@
+@import 'button';
 @import 'radio-block';
 @import 'colorpicker';
 @import 'checkbox';
