+++ src/styles/main.scss
@@ -25,6 +25,7 @@
   display: flex;
   @import 'components/main';
   @include font-smoothing;
+  font-family: var(--font-family-body);
   &__wrapper {
     display: flex;
     flex-grow: 1;
